Provide real containers to the transfer file dialog
