+++ b4n-tui/widgets/modal/selector.rs
@@ -30,7 +30,7 @@ pub struct Selector {
 
 impl Selector {
     /// Creates new [`Selector`] instance.
-    pub fn new(id: usize, caption: &'static str, options: &[&str], colors: &SelectModalColors) -> Self {
+    pub fn new<S: AsRef<str>>(id: usize, caption: &'static str, options: &[S], colors: &SelectModalColors) -> Self {
         let mut options = ActionsListBuilder::from_strings(options).build(None);
         options.highlight_first_item();
         let selected = options.get_highlighted_item_name().unwrap_or_default().to_owned();
