+++ b4n/ui/views/resources/mod.rs
@@ -4,4 +4,5 @@ pub use self::view::*;
 mod dialogs;
 mod menus;
 mod table;
+mod utils;
 mod view;
