+++ b4n/ui/views/resources/view.rs
@@ -201,8 +201,11 @@ impl ResourcesView {
 
     /// Shows transfer file dialog.
     pub fn ask_transfer_file(&mut self, is_download: bool) {
-        self.modal = new_transfer_dialog(is_download, &self.app_data, self.last_mouse_click.take());
-        self.modal.show();
+        if let Some(resource) = self.table.list.table.get_highlighted_resource() {
+            let tags = resource.data.as_ref().map(|d| d.tags.as_ref()).unwrap_or_default();
+            self.modal = new_transfer_dialog(is_download, &self.app_data, tags, self.last_mouse_click.take());
+            self.modal.show();
+        }
     }
 
     /// Displays a list of available contexts to choose from.
